shadows completed, minor changes for norminette

+++ inc/plane_intersection.h
@@ -6,7 +6,7 @@
 /*   By: irozhkov <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 16:36:31 by irozhkov          #+#    #+#             */
-/*   Updated: 2025/02/08 19:38:54 by irozhkov         ###   ########.fr       */
+/*   Updated: 2025/02/11 19:43:19 by irozhkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 # include "minirt.h"
 
 void	plane_intersection(t_scene *scene, t_item *item, t_ray *ray);
-int		plane_sh_intersection(t_item *item, t_ray *ray);
+double		plane_sh_intersection(t_item *item, t_ray *ray);
 
 #endif
